requirements.txt
Django==2.2
cymysql
django-cymysql==2.2
djangorestframework
django-guardian
bcrypt==3.2.0